+++ NEWS.md
@@ -33,6 +33,9 @@
 
 - Bug fix for `mix()` models as well as `iCov` models.
 
+- Single core thread for etTrans and rxSolve to speed up solving
+  without OpenMP
+
 # rxode2 5.1.2
 
 - `geom_cens()` / `stat_cens()` no longer emit "Ignoring unknown